changed doubles to ints. Also added javadoc comments to all methods in Player and Zombie classes.

+++ ZombieProjectModelClasses/src/zombieProject/Player.java
@@ -2,44 +2,76 @@ package zombieProject;
 
 public class Player extends Game{
 
-	private double x; //player's x coordinate
-	private double y; //player's y coordinate
+	private int x; //player's x coordinate
+	private int y; //player's y coordinate
 	
 	private double health; //player's health
 	
-	//constructor
-	public Player(double x, double y){
+	/**
+	 * player constructor
+	 * 
+	 * @param x initial x-coordinate
+	 * @param y initial y-coordinate
+	 */
+	public Player(int x, int y){
 		this.x = x;
 		this.y = y;
 		
 		health = 100;
-		
 	}
-
+	/**
+	 * returns player's x-coordinate
+	 */
 	public double getX(){
 		return x;
 	}
 	
-	public void setX(double x){
+	/**
+	 * set player's x-coordinate
+	 * 
+	 * @param x new x-coordinate
+	 */
+	public void setX(int x){
 		this.x = x;
 	}
 	
+	/**
+	 * returns player's y-coordinate
+	 */
 	public double getY(){
 		return y;
 	}
 	
-	public void setY(double y){
+	/**
+	 * set player's y-coordinate
+	 * 
+	 * @param y new y-coordinate 
+	 */
+	public void setY(int y){
 		this.y = y;
 	}
 	
+	/**
+	 * increase the players health
+	 * 
+	 * @param val the amount to be added to player's health
+	 */
 	public void increaseHealth(double val){
 		this.health += val;
 	}
 	
+	/**
+	 * decreases the player's health
+	 * 
+	 * @param val the amounte to be removed from player's health
+	 */
 	public void decreaseHealth(double val){
 		this.health -= val;
 	}
 	
+	/**
+	 * returns the player's health
+	 */
 	public double getHealth() {
 		return health;
 	}
