+++ ZombieProjectModelClasses/src/zombieProject/Zombie.java
@@ -2,34 +2,59 @@ package zombieProject;
 
 public class Zombie {
 
-		private double x; //zombie x-coordinate
-		private double y; //zombie y-coordinate
+		private int x; //zombie x-coordinate
+		private int y; //zombie y-coordinate
 		
 		private double health; //zombie health
 		
-		public Zombie(double x, double y){
+		/**
+		 * Zombie constructor
+		 * 
+		 * @param x initial x-coordinate
+		 * @param y initial y-coordinate
+		 */
+		public Zombie(int x, int y){
 			this.x = x;
 			this.y = y;
 			
 			health = 50;
 		}
 		
+		/**
+		 * returns the zombie's x-coordinate
+		 */
 		public double getX(){
 			return x;
 		}
 		
-		public void setX(double x){
+		/**
+		 * set the zombie's x-coordinate
+		 * 
+		 * @param x new x-coordinate
+		 */
+		public void setX(int x){
 			this.x = x;
 		}
 		
+		/**
+		 * returns the zombie's y-coordinate
+		 */
 		public double getY(){
 			return y;
 		}
 		
-		public void setY(double y){
+		/**
+		 * set the zombie's y-coordinate
+		 * 
+		 * @param y new y-coordinate
+		 */
+		public void setY(int y){
 			this.y = y;
 		}
 		
+		/**
+		 * return zombies health
+		 */
 		public double getHealth() {
 			return health;
 		}
@@ -38,7 +63,7 @@ public class Zombie {
 		 * bullet's coordinates are the same as the Zombie's.
 		 * In other words, Zombie got shot.
 		 * 
-		 * @param val
+		 * @param val the value to remove from health
 		 */
 		public void decreaseHealth(double val){
 			this.health -= val;
